Soften dependency version constraints to make VertFlow compatible with the latest version of Cloud Composer. Tidy docs. Move some utils around. Correct tests around logging.
