+++ src/operator.py
@@ -26,25 +26,25 @@ from airflow.utils.context import Context
 
 class VertFlowOperator(BaseOperator):
     def __init__(  # type: ignore
-            self,
-            project_id: str,
-            name: str,
-            image_address: str,
-            command: str,
-            arguments: list[str],
-            service_account_email_address: str,
-            co2_signal_api_key: str,
-            working_directory: str = "/",
-            port_number: int = 8080,
-            max_retries: int = 3,
-            timeout_seconds: int = 300,
-            initialisation_timeout_seconds: int = 60,
-            cpu_limit: int = 1,
-            environment_variables: dict = {},
-            memory_limit: str = "512Mi",
-            annotations: dict = {},
-            allowed_regions: Optional[Sequence[str]] = None,
-            **kwargs,
+        self,
+        project_id: str,
+        name: str,
+        image_address: str,
+        command: str,
+        arguments: list[str],
+        service_account_email_address: str,
+        co2_signal_api_key: str,
+        working_directory: str = "/",
+        port_number: int = 8080,
+        max_retries: int = 3,
+        timeout_seconds: int = 300,
+        initialisation_timeout_seconds: int = 60,
+        cpu_limit: int = 1,
+        environment_variables: dict = {},
+        memory_limit: str = "512Mi",
+        annotations: dict = {},
+        allowed_regions: Optional[Sequence[str]] = None,
+        **kwargs,
     ) -> None:
         """
         Execute a job in a Docker container on Cloud Run. Given a collection of allowed regions that the job can run in,
@@ -123,7 +123,9 @@ class VertFlowOperator(BaseOperator):
                                               `'-'     |_|                                      
         """
         logging.info(art)
-        logging.info("VertFlow is finding the greenest region to run your Cloud Run Job.")
+        logging.info(
+            "VertFlow is finding the greenest region to run your Cloud Run Job."
+        )
         cloud_run_regions = CloudRunRegions(self.project_id, self.co2_signal_api_key)
 
         try:
@@ -132,7 +134,7 @@ class VertFlowOperator(BaseOperator):
             logging.info(
                 f"Deploying Cloud Run Job {self.name} in {greenest['name']} ({greenest['id']}) "
                 f"where carbon intensity is {greenest['carbon_intensity']} gCO2eq/kWh. "
-                f"This is {closest['carbon_intensity'] - greenest['carbon_intensity']} gCO2eq/kWh lower than your closest region {closest['name']} ({closest['id']})."
+                f"This is {float(closest['carbon_intensity']) - float(greenest['carbon_intensity'])} gCO2eq/kWh lower than your closest region {closest['name']} ({closest['id']})."
             )
         except (ConnectionError, LookupError) as e:
             greenest = (
@@ -145,7 +147,7 @@ class VertFlowOperator(BaseOperator):
             )
 
         self.job = CloudRunJob(
-            greenest["id"],
+            str(greenest["id"]),
             self.project_id,
             self.name,
         )
