+++ test/build_on_cloud_composer/requirements.txt
@@ -1,0 +1,1 @@
+apache-airflow==2.2.5+composer
