+++ test/utils.py
@@ -1,6 +1,61 @@
+from time import time
+from typing import Optional
 from unittest import TestCase
 
-from src.utils import intersection_equal
+from src.utils import intersection_equal, wait_until
+
+
+class TestWaitUntil(TestCase):
+    def setUp(self) -> None:
+        self.number_of_calls_until_returns_true = 3
+        self.call_count = 0
+        self.wait_interval_seconds = 5
+
+    def condition_to_wait_for(self) -> bool:
+        if self.call_count < self.number_of_calls_until_returns_true:
+            self.call_count += 1
+            return False
+        else:
+            return True
+
+    def test_wait_until_blocks_until_condition_returns_true(self) -> None:
+
+        time_before_wait = time()
+
+        # Condition will return true after 15 seconds.
+        wait_until(self.condition_to_wait_for, 20, self.wait_interval_seconds)
+
+        # Validate that wait_until exits cleanly after c. 15 seconds.
+        self.assertTrue(
+            self.number_of_calls_until_returns_true * self.wait_interval_seconds
+            < time() - time_before_wait
+            <= self.number_of_calls_until_returns_true * self.wait_interval_seconds + 1
+        )
+
+    def test_wait_throws_if_timeout_is_reached_before_condition_returns_true(
+        self,
+    ) -> None:
+
+        time_before_wait = time()
+        timeout_seconds = 5
+        exception_thrown: Optional[Exception] = None
+
+        try:
+            wait_until(
+                self.condition_to_wait_for, timeout_seconds, self.wait_interval_seconds
+            )
+        except Exception as e:
+            exception_thrown = e
+
+        time_after_wait = time()
+
+        # Validate that wait_until exits after timeout
+        self.assertTrue(
+            timeout_seconds < time_after_wait - time_before_wait <= timeout_seconds + 1
+        )
+
+        # Validate that wait_until throws a TimeoutError
+        self.assertIsInstance(exception_thrown, TimeoutError)
 
 
 class TestIntersectionEqual(TestCase):
